+++ src/pages/validationRuleGroups/New.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { FormBase } from '../../components'
 import { DefaultNewFormContents } from '../../components/form/DefaultFormContents'
 import { SECTIONS_MAP, useOnSubmitNew } from '../../lib'
-import ValidationRuleGroupsFormFields from './form/ValidationRuleGroupsFormFields'
+import { ValidationRuleGroupsFormFields } from './form/ValidationRuleGroupsFormFields'
 import { initialValues, validate } from './form/validationRuleGroupsSchema'
 
 const section = SECTIONS_MAP.validationRuleGroup