+++ src/testUtils/builders.ts
@@ -68,6 +68,7 @@ import {
     validationRuleGroupsFormSchema,
     validationRuleGroupsListSchema,
 } from '../pages/validationRuleGroups/form/validationRuleGroupsSchema'
+import { validationRuleListSchema } from '../pages/validationRules/form/validationRuleSchema'
 import {
     CategoryMapping,
     DataElement,
@@ -508,9 +509,7 @@ export const testOrgUnit = (overwrites: Record<any, any> | undefined = {}) => {
     } as unknown as Partial<OrganisationUnit>
 }
 
-export const testValidationRuleGroups = (
-    overwrites: Record<any, any> = {}
-) => ({
+export const testValidationRuleGroup = (overwrites: Record<any, any> = {}) => ({
     ...generateMock(validationRuleGroupsListSchema, { mockeryMapper }),
     ...overwrites,
 })
@@ -519,6 +518,13 @@ export const testValidationRuleGroupsForm = (
     overwrites: Record<any, any> = {}
 ) => ({
     ...generateMock(validationRuleGroupsFormSchema, {
+        mockeryMapper,
+    }),
+    ...overwrites,
+})
+
+export const testValidationRule = (overwrites: Record<any, any> = {}) => ({
+    ...generateMock(validationRuleListSchema, {
         mockeryMapper,
     }),
     ...overwrites,
@@ -551,12 +557,3 @@ export const testOption = ({
         displayName: displayName ?? optionName,
     }
 }
-
-// TODO: change when schema for validationRule is available
-export const testValidationRule = ({
-    id = randomDhis2Id(),
-    displayName = faker.person.fullName(),
-} = {}) => ({
-    id,
-    displayName,
-})